// src/interfaces/sortBy.ts
enum SortBy{
    popularityAsc ='popularity.asc', 
    releaseDateAsc = 'release_date.asc',  
    revenueAsc = 'revenue.asc', 
    primaryReleaseDateAsc = 'primary_release_date.asc',
    originalTitleAsc ='original_title.asc',
    voteAverageAsc = 'vote_average.asc', 
    voteCountAsc = 'vote_count.asc',
    popularityDesc ='popularity.desc', 
    releaseDateDesc = 'release_date.desc',  
    revenueDesc = 'revenue.desc', 
    primaryReleaseDateDesc = 'primary_release_date.desc',
    originalTitleDesc ='original_title.desc',
    voteAverageDesc = 'vote_average.desc', 
    voteCountDesc = 'vote_count.desc'
}

export default SortBy;